#22790 Fix bug when uograde in msssql
