+++ dotCMS/src/integration-test/java/com/dotmarketing/startup/runonce/Task220824CreateDefaultVariantTest.java
@@ -29,7 +29,7 @@ public class Task220824CreateDefaultVariantTest {
     public static void prepare() throws Exception {
         IntegrationTestInitService.getInstance().init();
 
-        checkIfVariantColumnExist();
+        checkIfVariantDefaultExists();
     }
 
     private static void checkIfVariantDefaultExists() throws DotDataException {
@@ -39,34 +39,6 @@ public class Task220824CreateDefaultVariantTest {
 
         assertEquals("The DEFAULT Variant should exists", 1, results.size());
     }
-
-    private static void checkIfVariantColumnExist() throws SQLException {
-        final Connection connection = DbConnectionFactory.getConnection();
-
-        final ResultSet columnsMetaData = DotDatabaseMetaData
-                .getColumnsMetaData(connection, "contentlet_version_info");
-
-        boolean variantIdFound = false;
-        while(columnsMetaData.next()){
-            final String columnName = columnsMetaData.getString("COLUMN_NAME");
-            final String columnType = columnsMetaData.getString(6);
-
-            if ("variant_id".equals(columnName)) {
-                variantIdFound = true;
-
-                if (DbConnectionFactory.isPostgres()) {
-                    assertEquals("varchar", columnType);
-                } else {
-                    assertEquals("nvarchar", columnType);
-                }
-
-                assertEquals("The column sice should be 255", "255", columnsMetaData.getString(7));
-            }
-        }
-
-        assertTrue( variantIdFound, "Should exists de variant field in contentlet_version_info");
-    }
-
     /**
      * Method to test: {@link Task220824CreateDefaultVariant#executeUpgrade()}
      * when: the UT run
@@ -81,56 +53,6 @@ public class Task220824CreateDefaultVariantTest {
         upgradeTask.executeUpgrade();
 
         checkIfVariantDefaultExists();
-        checkIfVariantColumnExist();
-    }
-
-    /**
-     *  Method to test: {@link Task220824CreateDefaultVariant#executeUpgrade()}
-     *  When: Create a Contentlet before the {@link Task220824CreateDefaultVariant} run
-     *  Should: not going to have variant_id
-     *  When: Create a Contentlet after the {@link Task220824CreateDefaultVariant} run
-     *  Should:  have a variant_id equals to 1
-     *
-     * @throws DotDataException
-     */
-    @Test
-    public void conntentletVersionInfoWithVariant() throws DotDataException {
-        final ContentType contentType = new ContentTypeDataGen().nextPersisted();
-        final Contentlet contentletBefore = new ContentletDataGen(contentType).nextPersisted();
-
-        cleanAllBefore();
-
-        if (!DbConnectionFactory.isMsSql()) {
-            checkNotExistVariant(contentletBefore);
-        }
-
-        final Task220824CreateDefaultVariant upgradeTask = new Task220824CreateDefaultVariant();
-        upgradeTask.executeUpgrade();
-
-        final Contentlet contentletAfter = new ContentletDataGen(contentType).nextPersisted();
-
-        checkContentletVersionInfo(contentletBefore);
-        checkContentletVersionInfo(contentletAfter);
-    }
-
-    private void checkNotExistVariant(final Contentlet contentlet) throws DotDataException {
-        final ArrayList results = new DotConnect().setSQL(
-                        "SELECT * FROM contentlet_version_info WHERE identifier = ?")
-                .addParam(contentlet.getIdentifier())
-                .loadResults();
-
-        assertEquals(1, results.size());
-        assertNull(((Map) results.get(0)).get("variant_id"));
-    }
-
-    private void checkContentletVersionInfo(final Contentlet contentlet) throws DotDataException {
-        final ArrayList results = new DotConnect().setSQL(
-                        "SELECT * FROM contentlet_version_info WHERE identifier = ?")
-                .addParam(contentlet.getIdentifier())
-                .loadResults();
-
-        assertEquals(1, results.size());
-        assertEquals("DEFAULT", ((Map) results.get(0)).get("variant_id").toString());
     }
 
     @Test
@@ -154,6 +76,8 @@ public class Task220824CreateDefaultVariantTest {
         cleanAllBefore();
 
         final Task220824CreateDefaultVariant upgradeTask = new Task220824CreateDefaultVariant();
+
+        assertTrue(upgradeTask.forceRun());
         assertTrue(upgradeTask.forceRun());
 
         upgradeTask.executeUpgrade();
@@ -161,36 +85,12 @@ public class Task220824CreateDefaultVariantTest {
         assertFalse(upgradeTask.forceRun());
     }
 
-    private void cleanAllBefore() throws DotDataException {
+    private void cleanAllBefore() {
         final DotConnect dotConnect = new DotConnect();
 
         try {
             dotConnect.setSQL("DELETE FROM variant WHERE name = ?")
                     .addParam(VariantAPI.DEFAULT_VARIANT.name())
-                    .loadResult();
-        } catch (Exception e) {
-
-        }
-
-        try {
-            if (DbConnectionFactory.isMsSql()) {
-                final ArrayList<Map> loadResults = dotConnect.setSQL("SELECT name "
-                                + "FROM sysobjects so JOIN sysconstraints sc ON so.id = sc.constid "
-                                + "WHERE object_name(so.parent_obj) = 'contentlet_version_info' AND "
-                                + "sc.colid in  (select colid from syscolumns where name = 'variant_id')")
-                        .addParam(VariantAPI.DEFAULT_VARIANT.name())
-                        .loadResults();
-
-                dotConnect.setSQL("ALTER TABLE contentlet_version_info DROP CONSTRAINT " + loadResults.get(0).get("name").toString() )
-                        .loadResult();
-            }
-        } catch (Exception e) {
-
-        }
-
-        try {
-            dotConnect
-                    .setSQL("ALTER TABLE contentlet_version_info DROP COLUMN variant_id")
                     .loadResult();
         } catch (Exception e) {
